fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pydantic
python-jose
paramiko